\name{track_OCP}
\alias{track_OCP}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Clustering Behaviour along NAM }
\description{
  Evaluates the overall clustering performance of sub-networks along removal
  process.
}
\usage{
  track_OCP(histrval, msym)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{histrval}{ Integer vector. Index of last sub-network of node membership. }
  \item{msym}{ A sympatry matrix.}
}
\details{
  This function is auxilliary and is not intended to be called by the user.
}
\author{ Daniel A. Dos Santos }
\seealso{ 
  \code{\link{nam}} function makes a call to this function.
}
\keyword{ methods }

