\name{viewsp}
\alias{viewsp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Spatial Expressions of NAM Classes }
\description{
  Displays marked set of records of a NAM partition applied on
  a given data set. 
}
\usage{viewsp()}
%- maybe also 'usage' for other objects documented here.
\arguments{
  There is no argument to be invoked. 
}
\details{
  This function is auxilliary and is not intended to be called by the user. It 
  creates a GUI interface to see distributions in a geographic context.
}
\author{Daniel A. Dos Santos <dadossantos@csnat.unt.edu.ar>}
\seealso{ 
  \code{\link{cleavogram}} calls this function.
}
\keyword{internal}

