\name{sciobius2x2}
\alias{sciobius2x2}
\docType{data}
\title{Sciobius Grid Occupancy Data}
\description{
  Presence/absence data of 47 South African species of weevils over a squared 
  grid of 2 x 2 degrees.
}
\usage{data(sciobius2x2)}
\format{
  A data frame with 49 observations on 25 columns. Columns are the grid cells in the
  study area (South Africa). The first pair of rows correspond to the coordinates of cells
  in the grid system (i.e. numeric indices of the cells in the rectangular layout of the grid).

  The rest of rows are the species. Presence is indicated with 1 and absence with 0.
}
\details{
 Morrone (1994) provides the original matrix for this classical example, but ulterior
 corrections are introduced by Mast and Nyffeler (2003). The corrected version of 
 the data are here represented. Species names are coded through two letters following the
 nomclature of Mast and Nyffeler (2003). 
}
\source{
  Schoeman, A. S. 1983. \emph{Revision of the Genus Sciobius Schonherr (Coleoptera: Curculionidae).}
  Entomol. Mem. Dep. Agric. Repub. S. Afr. 59:1-50.
}
\references{
  Morrone, J. J. 1994. \emph{On the Identification of Areas of Endemism.} Syst. Biol. 43:438-441.

  Mast, A. R., and R. Nyffeler. 2003. \emph{Using a Null Model to Recognize Significant Co-
  occurrence Prior to Identifying Candidate Areas of Endemism.} Syst. Biol. 52:271-280.
}
\keyword{datasets}
